refactor: break down ffi_primitives into modular components (Issue #22 Phase 2)
Create proper Rust module structure for FFI primitives by decomposing the
monolithic ffi_primitives.rs into logical submodules within src/ffi/primitives/:

New modules:
- context.rs: VM context management
- types.rs: C type parsing utilities
- library.rs: Library loading and management
- calling.rs: C function calling primitives
- enums.rs: Header parsing and enum support
- callbacks.rs: C function pointer callbacks
- memory.rs: Memory management and safety checking
- wrappers.rs: Context-aware wrapper functions

Changes:
- Create src/ffi/primitives/ directory with 8 focused modules
- Each module exports public functions and maintains clear responsibilities
- mod.rs centralizes exports and maintains API compatibility
- Remove old src/ffi_primitives.rs file
- Update src/lib.rs to import from ffi module
- Update src/ffi/mod.rs to declare primitives submodule

Testing:
- All 809 tests pass (no regressions)
- Zero clippy warnings
- Code compiles cleanly

This completes Issue #22 technical debt item #1 in Phase 2 with true module
decomposition. The structure improves code organization, testability, and
maintainability while preserving API compatibility.
